HOLO-66 <chore: 안쓰는 폴더 삭제>
