+++ src/app/sign-in/_types/userType.ts
@@ -1,7 +1,9 @@
-import { Tables } from "@/lib/types/supabase";
+//이것도 기능테스트후 삭제예정
+
+// import { Tables } from "@/lib/types/supabase";
 
 
-export type User = Tables<'users'> ;
+// export type User = Tables<'users'> ;
