+++ src/app/sign-up/_types/auth.ts
@@ -1,15 +1,16 @@
+// 기능 테스트후 삭제예정
 
-import { User } from "@/app/sign-up/_types/userType";
+// import { User } from "@/app/sign-up/_types/userType";
 
-export type Session = {
-  accessToken: string;
-  refreshToken: string;
-};
+// export type Session = {
+//   accessToken: string;
+//   refreshToken: string;
+// };
 
-export type AuthState = {
-  isLoggedIn: boolean;
-  user: User | null;
-  session: Session | null;
-  setAuth: (user: User | null, session: Session | null) => void;
-  clearAuth: () => void;
-};
+// export type AuthState = {
+//   isLoggedIn: boolean;
+//   user: User | null;
+//   session: Session | null;
+//   setAuth: (user: User | null, session: Session | null) => void;
+//   clearAuth: () => void;
+// };
